requirements.txt
numpy==1.24.3
opencv-python==4.8.0.74
onnx==1.14.1
onnxruntime-gpu==1.16.0
insightface==0.7.3
psutil==5.9.5
pillow==10.0.0
torch==2.1.0
torchvision==0.16.0
tensorflow==2.13.0
opennsfw2==0.10.2
protobuf==4.24.0
tqdm==4.66.0
piexif
lpips
numba>=0.57.0
scipy>=1.11.0
basicsr>=1.4.2
realesrgan>=0.3.0

